Move some methods into decks/counts.rs
